Check permissions before accepting and rejecting invites. Add accept and reject routes for topic invites.

+++ src/main/java/saturday/controllers/TopicMemberController.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.*;
 import saturday.domain.Entity;
 import saturday.domain.Topic;
@@ -14,6 +15,7 @@ import saturday.domain.TopicMember;
 import saturday.domain.TopicMemberRequest;
 import saturday.exceptions.TopicMemberNotFoundException;
 import saturday.services.EntityService;
+import saturday.services.PermissionService;
 import saturday.services.TopicMemberService;
 import saturday.services.TopicService;
 
@@ -22,14 +24,16 @@ public class TopicMemberController {
     private final TopicMemberService topicMemberService;
     private final EntityService entityService;
     private final TopicService topicService;
+    private final PermissionService permissionService;
 
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @Autowired
-    public TopicMemberController(TopicMemberService topicMemberService, EntityService entityService, TopicService topicService) {
+    public TopicMemberController(TopicMemberService topicMemberService, EntityService entityService, TopicService topicService, PermissionService permissionService) {
         this.topicMemberService = topicMemberService;
         this.entityService = entityService;
         this.topicService = topicService;
+        this.permissionService = permissionService;
     }
 
     @RequestMapping(value = "/topic_members/{id}", method = RequestMethod.GET)
@@ -48,6 +52,10 @@ public class TopicMemberController {
 
     @RequestMapping(value = "/topic_members", method = RequestMethod.POST)
     public ResponseEntity<TopicMember> saveTopicMember(@RequestBody TopicMemberRequest topicMemberRequest) throws BadHttpRequest {
+        if(!permissionService.canAccess(topicMemberRequest)) {
+            throw new AccessDeniedException("Authenticated entity does not have sufficient permissions.");
+        }
+
         Topic topic = topicService.findTopicById(topicMemberRequest.getTopicId());
         if(topic == null) {
             throw new BadHttpRequest(new Exception("Invalid topic id " + topicMemberRequest.getTopicId()));
@@ -76,11 +84,16 @@ public class TopicMemberController {
 
     @RequestMapping(value = "/topic_members/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<TopicMember> delete(@PathVariable(value = "id") int id) {
+        if(!entityService.getAuthenticatedEntity().isAdmin()) {
+            throw new AccessDeniedException("Authenticated entity does not have sufficient permissions.");
+        }
+
         try {
             topicMemberService.delete(id);
         } catch(EmptyResultDataAccessException ex) {
             throw new TopicMemberNotFoundException("No topic member with id " + id + " exists!");
         }
+
         return new ResponseEntity<>(HttpStatus.OK);
     }
 }
